+++ Trees/Structs.h
@@ -1,21 +1,21 @@
 #pragma once
 
-struct Point2d
+struct Point2f
 {
-	double x, y;
+	float x, y;
 };
 
-struct Point3d
+struct Point3f
 {
-	double x, y, z;
-	Point3d () {}
-	Point3d (double x, double y, double z)
+	float x, y, z;
+	Point3f () {}
+	Point3f (float x, float y, float z)
 	{
 		this->x = x;
 		this->y = y;
 		this->z = z;
 	}
-	Point3d (std::string s)
+	Point3f (std::string s)
 	{
 		x = atof (substring (s, 0, s.find_first_of (' ') - 1).c_str ());
 		y = atof (substring (s, s.find_first_of (' ') + 1, s.find_last_of (' ') - 1).c_str ());
