+++ Trees/Trees.cpp
@@ -8,11 +8,11 @@
 //Read about Poisson disk sampling and dart throwing -- simple idea to generate visually pleasing distribution patterns. Also think about using Autodesk 123D to capture real tree mdels
 using namespace std;
 
-bool is_good (vector<vector<Point2d>> grid, Point2d pt, double min_dist, double cell_size)
+bool is_good (vector<vector<Point2f>> grid, Point2f pt, double min_dist, double cell_size)
 {
 	int x = ceil (pt.x / cell_size), y = ceil (pt.y / cell_size), w = grid.size (), h = grid[0].size ();
 	double dist;
-	Point2d tmp;
+	Point2f tmp;
 	for (int i = x - 2; i <= x + 2 && i < w; i++) {
 		for (int j = y - 2; j <= y + 2 && j < h; j++) {
 			if (i < 0 || j < 0) continue;
@@ -27,15 +27,15 @@ bool is_good (vector<vector<Point2d>> grid, Point2d pt, double min_dist, double 
 
 void Trees::generateTrees (double width, double height, double min_dist, int num, int age)
 {
-	trees.push_back (pair<Point2d, Tree> (Point2d (), tree_objs[0]));
+	trees.push_back (pair<Point2f, Tree> (Point2f (), tree_objs[0]));
 	return;
 
 	double cell_size = min_dist / sqrt (2), gwidth = width / cell_size, gheight = height / cell_size, r, angle;
-	auto grid = vector<vector<Point2d>> (gwidth, vector<Point2d> (gheight));
-	auto worklist = deque<Point2d> ();
+	auto grid = vector<vector<Point2f>> (gwidth, vector<Point2f> (gheight));
+	auto worklist = deque<Point2f> ();
 	default_random_engine gen;
 	uniform_real_distribution<double> r_dist (0, 2), angle_dist (0, 2 * M_PI), dist (0.0, 1.0);
-	Point2d pt = Point2d { dist (gen)*width, dist (gen)*height }, new_pt;
+	Point2f pt = Point2f { dist (gen)*width, dist (gen)*height }, new_pt;
 
 	worklist.push_back (pt);
 	trees.push_back (make_pair (pt, tree_objs[dist (gen)*tree_objs.size ()]));
@@ -48,7 +48,7 @@ void Trees::generateTrees (double width, double height, double min_dist, int num
 		for (size_t i = 0; i < num; i++) {
 			r = r_dist (gen);
 			angle = angle_dist (gen);
-			new_pt = Point2d { pt.x + r*cos (angle), pt.y + r*sin (angle) };
+			new_pt = Point2f { pt.x + r*cos (angle), pt.y + r*sin (angle) };
 			if (new_pt.x > 0 && new_pt.x < width &&
 				 new_pt.y > 0 && new_pt.y < height &&
 				 is_good (grid, new_pt, min_dist, cell_size)) {
